+++ src/material/material.h
@@ -12,10 +12,10 @@ vec3 reflect(const vec3 &v, const vec3 &n) {
 }
 
 ///refraction
-bool refract(const vec3 &v, const vec3 &n, float ni_over_nt, vec3 &refracted) {
+bool refract(const vec3 &v, const vec3 &n, double ni_over_nt, vec3 &refracted) {
   vec3 uv = unit_vector(v);
-  float dt = dot(uv, n);
-  float discriminant = 1.0 - ni_over_nt * ni_over_nt * (1 - dt * dt);
+  double dt = dot(uv, n);
+  double discriminant = 1.0 - ni_over_nt * ni_over_nt * (1 - dt * dt);
   if (discriminant > 0) {
     refracted = ni_over_nt * (uv - n * dt) - n * sqrt(discriminant);
     return true;
@@ -24,8 +24,8 @@ bool refract(const vec3 &v, const vec3 &n, float ni_over_nt, vec3 &refracted) {
 }
 
 /// Schlick Approximation
-float schlick(float cosine, float ref_idx) {
-  float r0 = (1 - ref_idx) / (1 + ref_idx);
+double schlick(double cosine, double ref_idx) {
+  double r0 = (1 - ref_idx) / (1 + ref_idx);
   r0 = r0 * r0;
   return r0 + (1 - r0) * pow((1 - cosine), 5);
 }
@@ -62,7 +62,7 @@ class lambertian : public material {
 class metal : public material {
  public:
   metal(const vec3 &a) : albedo(a) {}
-  metal(const vec3 &a, float f) : albedo(a) { if (f < 1) fuzz = f; else fuzz = 1; }
+  metal(const vec3 &a, double f) : albedo(a) { if (f < 1) fuzz = f; else fuzz = 1; }
   virtual bool scatter(const ray &r_in, const hit_record &rec, vec3 &attenuation, ray &scattered) const {
     vec3 reflected = reflect(unit_vector(r_in.direction()), rec.normal);
     scattered = ray(rec.p, reflected + fuzz * random_in_unit_sphere());
@@ -70,21 +70,21 @@ class metal : public material {
     return (dot(scattered.direction(), rec.normal) > 0);
   }
   vec3 albedo;
-  float fuzz;
+  double fuzz;
 };
 
 ///dielectric
 class dielectric : public material {
  public:
-  dielectric(float ri) : ref_idx(ri) {}
+  dielectric(double ri) : ref_idx(ri) {}
   virtual bool scatter(const ray &r_in, const hit_record &rec, vec3 &attenuation, ray &scattered) const {
     vec3 outward_normal;
     vec3 reflected = reflect(r_in.direction(), rec.normal);
-    float ni_over_nt;
+    double ni_over_nt;
     attenuation = vec3(1.0, 1.0, 1.0);
     vec3 refracted;
-    float reflect_prob;
-    float cosine;
+    double reflect_prob;
+    double cosine;
     if (dot(r_in.direction(), rec.normal) > 0) {
       outward_normal = -rec.normal;
       ni_over_nt = ref_idx;
@@ -107,6 +107,6 @@ class dielectric : public material {
     return true;
   }
 
-  float ref_idx;
+  double ref_idx;
 };
 #endif //RAY_MATERIAL_MATERIAL_H_
