+++ src/objects/sphere.h
@@ -9,12 +9,12 @@
 class sphere : public hitable {
  public:
   sphere();
-  sphere(vec3 cen, float r) : center(cen), radius(r) {};
-  sphere(vec3 cen, float r, shared_ptr<material> m) : center(cen), radius(r), mat_ptr(m) {};
-  virtual bool hit(const ray &r, float tmin, float tmax, hit_record &rec) const override;
-  virtual bool bounding_box(float time0, float time1, aabb &box) const override;
+  sphere(vec3 cen, double r) : center(cen), radius(r) {};
+  sphere(vec3 cen, double r, shared_ptr<material> m) : center(cen), radius(r), mat_ptr(m) {};
+  virtual bool hit(const ray &r, double tmin, double tmax, hit_record &rec) const override;
+  virtual bool bounding_box(double time0, double time1, aabb &box) const override;
   vec3 center;
-  float radius;
+  double radius;
   shared_ptr<material> mat_ptr;
 
  private:
@@ -27,14 +27,14 @@ class sphere : public hitable {
   }
 };
 
-bool sphere::hit(const ray &r, float t_min, float t_max, hit_record &rec) const {
+bool sphere::hit(const ray &r, double t_min, double t_max, hit_record &rec) const {
   vec3 oc = r.origin() - center;
-  float a = dot(r.direction(), r.direction());
-  float b = dot(oc, r.direction());
-  float c = dot(oc, oc) - radius * radius;
-  float discriminant = b * b - a * c;
+  double a = dot(r.direction(), r.direction());
+  double b = dot(oc, r.direction());
+  double c = dot(oc, oc) - radius * radius;
+  double discriminant = b * b - a * c;
   if (discriminant > 0) {
-    float temp = (-b - sqrt(b * b - a * c)) / a;
+    double temp = (-b - sqrt(b * b - a * c)) / a;
     if (t_min < temp && temp < t_max) {
       rec.t = temp;
       rec.p = r.point_at_parameter(rec.t);
@@ -59,7 +59,7 @@ bool sphere::hit(const ray &r, float t_min, float t_max, hit_record &rec) const 
   return false;
 }
 
-bool sphere::bounding_box(float t0, float t1, aabb &box) const {
+bool sphere::bounding_box(double t0, double t1, aabb &box) const {
   vec3 scale = vec3(radius, radius, radius);
   box = aabb(center - scale, center + scale);
   return true;
@@ -70,37 +70,37 @@ class moving_sphere : public hitable {
  public:
   moving_sphere() {}
   moving_sphere(vec3 cen0, vec3 cen1,
-                float t0, float t1,
-                float r,
+                double t0, double t1,
+                double r,
                 shared_ptr<material> m) :
       center0(cen0), center1(cen1),
       time0(t0), time1(t1),
       radius(r),
       mat_ptr(m) {};
-  virtual bool hit(const ray &r, float tmin, float tmax, hit_record &rec) const override;
-  virtual bool bounding_box(float t0, float t1, aabb &box) const override;
-  vec3 center(float time) const;
+  virtual bool hit(const ray &r, double tmin, double tmax, hit_record &rec) const override;
+  virtual bool bounding_box(double t0, double t1, aabb &box) const override;
+  vec3 center(double time) const;
   vec3 center0, center1;
-  float time0, time1;
-  float radius;
+  double time0, time1;
+  double radius;
   shared_ptr<material> mat_ptr;
 };
 
 /// time0 から time1の間で移動する球の
 /// 時刻 t における中心座標
-vec3 moving_sphere::center(float time) const {
+vec3 moving_sphere::center(double time) const {
   return center0 + ((time - time0) / (time1 - time0)) * (center1 - center0);
 }
 
 /// 交差判定
-bool moving_sphere::hit(const ray &r, float t_min, float t_max, hit_record &rec) const {
+bool moving_sphere::hit(const ray &r, double t_min, double t_max, hit_record &rec) const {
   vec3 oc = r.origin() - center(r.time());
-  float a = dot(r.direction(), r.direction());
-  float b = dot(oc, r.direction());
-  float c = dot(oc, oc) - radius * radius;
-  float discriminant = b * b - a * c;
+  double a = dot(r.direction(), r.direction());
+  double b = dot(oc, r.direction());
+  double c = dot(oc, oc) - radius * radius;
+  double discriminant = b * b - a * c;
   if (discriminant > 0) {
-    float temp = (-b - sqrt(discriminant)) / a;
+    double temp = (-b - sqrt(discriminant)) / a;
     if (t_min < temp && temp < t_max) {
       rec.t = temp;
       rec.p = r.point_at_parameter(rec.t);
@@ -121,7 +121,7 @@ bool moving_sphere::hit(const ray &r, float t_min, float t_max, hit_record &rec)
   return false;
 }
 
-bool moving_sphere::bounding_box(float t0, float t1, aabb &box) const {
+bool moving_sphere::bounding_box(double t0, double t1, aabb &box) const {
   vec3 scale = vec3(radius, radius, radius);
   aabb box0(center(t0) - scale, center(t0) + scale);
   aabb box1(center(t1) - scale, center(t1) + scale);
