+++ src/utils/hitable.h
@@ -10,7 +10,7 @@
 class material;
 
 struct hit_record {
-  float t;
+  double t;
   vec3 p;
   vec3 normal;
   shared_ptr<material> mat_ptr;
@@ -26,8 +26,8 @@ struct hit_record {
 
 class hitable {
  public:
-  virtual bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const = 0;
-  virtual bool bounding_box(float t0, float t1, aabb &box) const = 0;
+  virtual bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const = 0;
+  virtual bool bounding_box(double t0, double t1, aabb &box) const = 0;
 };
 
 #endif //RAY_UTILS_HITABLE_H_
