+++ src/utils/ray.h
@@ -9,19 +9,19 @@
 class ray {
  public:
   ray() {}
-  ray(const vec3& a, const vec3& b, float time = 0.0f) {
+  ray(const vec3 &a, const vec3 &b, double time = 0.0) {
     A = a;
     B = b;
     _time = time;
   }
   vec3 origin() const { return A; }
   vec3 direction() const { return B; }
-  float time() const { return  _time; }
-  vec3 point_at_parameter(float t) const { return A + t * B; }
+  double time() const { return _time; }
+  vec3 point_at_parameter(double t) const { return A + t * B; }
 
   vec3 A;
   vec3 B;
-  float _time;
+  double _time;
 };
 
 #endif //RAY_RAY_H
